adr-literature-summarizer/requirements.txt
streamlit
spacy
PyPDF2
python-docx
pandas
